Move background_helper to two_sphere_model
